Create feature_request.md
This file will show a template that users can use to submit issues on GitHub, specifically to ask for enhancements. Using a template assures that users respect issue guidelines and hopefully makes the issue of better quality for the developer. This file can be edited to be more custom for the AcaciaLinux project, as this is just the basic GitHub template